Split/Merge High Density Boxes for Load Balancing
Previously we have introduced Box splitting to improve load balancing. This
behavior is controlled by three parameters:
```
    warpx.split_high_density_boxes=false`
    warpx.split_high_density_boxes_threshold=1.1`
    warpx.split_high_density_boxes_min_box_size=8`
```

In this PR, we have added supported for merging Boxes as well. Merging is
controlled by the same runtime parameters. A merge is performed when the
merged Boxes have a load below the threshold and a size below the max grid
size.